Implement new fragment-based Preferences system
This is an overhaul of the whole system, using fragments instead
of PreferenceScreens. It has a dynamic single/dual-pane system that
works similarly to preference headers, except you can use normal
preference layouts instead of only submenu headers.

Each subscreen is defined by its own fragment - the system includes
some basic behaviour baked in (automatic summary handling, launching
new fragments on a click etc.) and is easily extensible if you want
to add custom functionality for things like updates and listeners.

It uses the app's theme and updates immediately when the user
picks a new one, and also the SwitchPreferences are finally here!