Remove extraneous refresh calls in threadlist/forumlist loaders
I'm not sure if these are fixing some edge case, but they just seem to keep triggering
loader restarts every time a loader loads something. So there's a restart, some data has
changed (which the loader uses to update the display contents) and that triggers yet
another restart on the same data, which probably hasn't changed in those few milliseconds.

Maybe there's a reason for it which is why I made a separate commit, but yeah