Rework thread rating handling and 'fix' the Film Dump
I've broken out the ratings handling code into a new utility class
that handles all the parsing and drawables, and wrangles the different
types of ratings on the site. Hopefully it'll be easier to add to if
this crops up again or if they change the URLs for the different rating
types.

It also lets you check the type of rating you're looking at, so things
can be handled in code if necessary - and it's doing that right now for
the amazing 'make the empty thread icon view visible' hack when you're
looking at film dump threads. It's a stopgap until... someone makes it
look nicer, but it works and hopefully it'll make it easier to manage
the layouts dynamically.

Also I added the sweet star icons from the site, feel free to replace
the hell out of them