Move to general ElementCollection class, use it for ratings+secondary tags
Added secondary tag handling, and moved the shared code into a
general ElementCollection class where this stuff can be listed,
drawables retrieved and so on.

Secondary tags are a little decoration on the thread tags, I
didn't spend much time on it. Also the database code needs checking,
it was a little crazy in there and I just wanted to add a new
int field and a column to hold it. It seems to work fine but I was
winging it, so a check from someone who knows it would be good!