Switch nav drawer to support lib version
I've reworked the navigation stuff to use the new Design library, it's a bit easier
to work with too (mostly defined in XML with a callback for item selection).
I added a material-ish header without an actual graphic (just a horrible blown-up
app icon with a tint because why not), and a mini avatar icon. Pre-lollipop will
need some library or custom code to make it round.

Forum/thread names update immediately now (well, as soon as the loader is done)
I had to make a choice about exactly what happens with this stuff, since you
can have a thread open (with an associated forum) and then visit a different
forum in its fragment without opening a thread. So you end up with a forum
and a thread that are unconnected, so which do you show in the menu?

I went with the last fragment to update getting to set the details - it's all handled
in that setNavIds method, so it could be tweaked with a call to that whenever a
fragment is in view, or something. The loaders are kind of all over the place so it's
still inconsistent, things triggering in the background and re-updating 'what you're
looking at', but that's always been an issue!