Treat missing thread IDs as an error in PostReplyFragment (avoids NPEs)
Some people are getting a crash in SendPostRequest because they're expecting a thread ID to exist,
and apparently it sometimes doesn't! Not a good situation all round!
I've added some checks that back out when it spots the thread ID is missing, but I'm not totally sure
exactly when these situations would occur - maybe there's a more elegant way to recover from them?