Rework sharedprefs to use type-validated refs instead of string literals
I've centralised all the preference keys as string resources, so everything
can refer to a single variable instead of using string literals. I didn't like
how brittle it was to have copies of the same string that are at risk of typos
and hard to refactor, and the way the rest of the app needed to know the internals
of how preferences are handled.

The preference keys are stored as resource strings (so the settings XML can
refer to them), and there's a set of constants in the Keys file which the app
can use. Those are passed to set and get methods in AwfulPreferences, and they're
grouped in typedefs according to the type of preference they set (int, boolean etc).
So you can set any preference by its key constant, and passing the correct type is
enforced by the typedef-annotated method.

Hopefully this'll avoid any future bugs where someone accidentally mistypes something
or uses the wrong type and then who knows what's going on in there