Notification for PMs
Added a basic PM manager class that parses the UCP page and looks for a new message alert.
It calls back to the main activity so it can pop up a snackbar with a link for the PM activity intent.

I also changed stuff to use the real UCP url constant, it's necessary for getting PMs from
the bookmark list network request, but it probably should be used everywhere in general? Seems to
 work ok so far.