Add BBcode option for text select/insert action menu
Added a BBcode option to the action menu since older Android versions use that when text is selected,
and it covers up the action bar with the original BBcode button.
Put in a workaround for this action mode deselecting text when you select an option.
Also added the BBcode option to the insert action menu for consistency (only appears on newer APIs)

Fixed an issue on older devices where selected text remained selected after being replaced