Simplify database upgrade path, fix SQL crashes
Removed the cases where the same table is dropped further down, making the earlier cases redundant.
Also fixes the crashes where a table is recreated and later ALTERed to add a column, but the recreation
uses the most recent table structure which already includes that column. Which will help that one guy
who's been trying to start the app 600+ times