Create a themed bottom sheet component, refactor code to use it
Made a more generic version of the BottomSheetDialog that handles theming and stuff, to keep all
the code in one place.
Reworked MessageFragment to use this - after fetching and creating the list of PostIcons, it uses them
to initialise the ThemedBottomSheetDialog and holds a reference to that instead.
Also refactored MessageComposer to use it.