Implement Announcements
Added some components to parse thread lists for announcements, keep a record of which
have been seen and viewed (for notifications and unread counts), and display
the current announcements in a fragment.

This is basically the same idea as the PM manager - thread lists are where announcements
are listed, so whenever a thread list is parsed, run it through the announcement manager
so it can pull out the necessary info. The manager informs listeners when something
important happens, e.g. first live update this session, or whenever the announcement list
changes. It also has methods to get the current unread count, display the current announcements,
and mark them all as read.

The announcements page currently reuses AwfulPosts and the thread HTML (even though
announcements are structurally different), and loads them directly from the site without
caching any in the DB. Hopefully this works well enough!

Also did some refactoring to make code reuse easier