Remove legacy WebView configs, move code into AwfulWebView
Since everything works with the config from ThreadDisplayFragment, I've removed
the old config methods and moved the thread stuff into AwfulWebView itself as an
automatic initialisation step.

I've added some basic utility methods to AwfulWebView so calls can be simpler,
and the implementation details are hidden in a single place.