Stop page parsing from directly updating forum DB
ForumsRepository should be responsible for handling any changes to the forums DB.
The forum parsing code in AwfulForum was scraping data and rewriting the database entries,
removing some important data and messing things up.

What complicates things is the INDEX field - it's set when the forum list is refreshed, and used to
maintain ordering for forums and their subforums, so they can be grouped together and stay in
the same place. Inserting new forums means the order needs to be rewritten, so it's better to just
do that as part of a full refresh instead of shuffling records around.

This means that subforum titles aren't being updated every time their parent forum is parsed,
which was pretty inconsistent anyway, so I've bumped the full refresh to daily as a compromise