Add app state preferences for storing misc data
I've added a new store for things like update timestamps and the like, accessible through
the AwfulApplication singleton.

Forum updates were using the default shared prefs to store the last update time, but those automatically call
onPreferenceChanged and that triggers all kinds of updates. It meant the timestamp needed to be set
*after* the data was updated, so the forum list would refresh after the changes were written, which is really
bad and brittle. It also meant multiple updates for no good reason.

So sticking the data in a separate prefs store eliminates that and makes it easier to control when things update.
I've put the known/read announcements state in there too since that's along the same lines, app not user data.