Fix mobile twitters, add automatic dark theming
Updated the tweet embedding regex to catch arbitrary subdomains.

I've added some code to AwfulTheme to determine whether the theme is 'dark' or not, basically
checking if inherits from a Light theme or if it's a custom theme marked as dark. This is used in
generating the thread HTML to stick a 'dark-theme' data attribute on the CSS link, and then the
tweet embedding JS can just check that to determine if it's a dark theme or not - no need to hardcode.

That fixes things like the OLED theme not using dark twitter themes - this can probably be used elsewhere!

I also tweaked the javascript a little to try and make it a little more efficient (compiled the regex and made
the tweet-loading calls act on individual tweets)