Rework posting flow and draft handling
I've changed the way posting works, removing menu options and pushing them into modal dialogs instead.
So 'save draft' is now an option when the user leaves the reply activity with an unsent post, and 'preview'
is an option on the 'submit post' confirmation dialog.

The 'discard draft?' prompt is meant to force the decision to happen when the user is leaving the reply window,
when they know whether they want to keep it or save it for later. The previous behaviour saved it automatically
and then asked whether the user wanted to keep it when they next replied to that thread, which could be
months later (if ever). So this way garbage can go straight where it belongs

The 'restore draft?' prompt is no longer dismissable, so the user has to make a decision. Before you could
accidentally cancel the dialog and it would default to discarding the draft.
Having it default to using the draft is another option, I think a forced decision is better though

The 'submit post?' dialog now has the preview function as a third choice. Mostly to make it more discoverable
and encourage people to use it, plus it fits with the main site's POST / PREVIEW button layout
(which isn't necessarily an exemplar or anything I know ok)

Automatic draft saving now happens in #onPause, should hopefully cover more situations where the reply
can be 'lost'. I've moved it into the general cleanup method, and removed some of the calls that are basically
covered by onPause (like the viewpager hiding the reply page)

Also various code tweaks while I was in there, some documentation, and I've ditched the reply content state
variable so we can just grab the current contents of the message composer instead