Check for PMs and announcements on sync
Added checks for those so they can notify the user on app start.
They're handled whenever the user loads a page with one of those notification links -
the bookmarks page for PMs, and an actual forum page (with the thread list) for announcements.

That's not great if the user doesn't visit one of those pages, so now those pages are loaded on sync
to trigger an update