Rework draft logic, add error messages when activity cannot start
I've moved the draft code around so the loader handles loading any data, and creates a SavedDraft object
encapsulating all the related variables. A separate method gets called when the reply data is loaded, checks
if there's any SavedDraft object to handle, and then decides whether to give the user the option to use it.

The idea is to separate the deserialisation from the general workflow, and have the decision logic pulled out
into a clearer and more explicit flow, and make the overall state a bit simpler.