Fix ThemeSettings crash without storage permissions, remove placeholder ListPreference entries
There was an issue where list preferences that need to be dynamically generated (themes and layouts)
weren't being initialised if there was a permission request that was also denied. This left the preference
in a state where the entries and values didn't match, so you could get an invalid index

I've cleaned all that up and made it so the preferences are always initialised and get refreshed if there's
a permission change. I've also added a dialog to tell people why they need it, with the frog staring at them

Since the font and theme ListPreferences are populated entirely programmatically, I've deleted the XML
data for their entries that gets completely discarded anyway.

Fixes Crashlytics issue #699