Add Crashlytics logging for issue #717, and Crashlytics init flag
Not sure what's causing that, shouldn't be in a paused state when that method is called. It's possible
though since it's coming from the webview's link handler...

I've also added a flag to check if Crashlytics has been initialised - before it was just a 'not DEBUG' check
which is a bit brittle (also Crashlytics is enabled in debug after X hours anyway), and made things use it.