Fix AwfulProvider#query method not handling bad data (Crashlytics #719)
Sometimes when the ContentProvider is queried (through a CursorLoader
usually) bad data is passed, and the #query method has no fallback to
handle it. E.g. the PM system sometimes passes a PM ID of -1, the
UriMatcher doesn't catch it (probably because the number wildcard doesn't
match the -) and it never matches a Uri type. It still makes the query
without setting things like the tables parameter, and then it throws.

I've made the query method return null for bad data, and the querying
methods can be responsible for handling it (especially when they're
passing bad data in the first place)

The other methods (#update etc) don't have any proper validation either,
so bad data could end up being passed in and cause problems. The whole
thing needs reworking to be honest