Add 'standarderror' check to AwfulError's parsing, handles rate limiting
This 'standarderror' class appears in some error pages, PostRequest was
explicitly handling it in its #handleResponse method and throwing an
ACCESS DENIED AwfulError. The same class appears in the 'you're posting
too fast' error pages too, and possibly in some other situations?

So I've pulled that check out and put it in the normal AwfulError parsing
method, so it can run on everything and throw an error just like
everything else does.