Rework thread context menus
I've split the config/handler code for the two menus into their own
classes, inheriting from a basic menu fragment. Now each menu's config
and action handling is defined in its own subclass, which will hopefully
be easier to maintain and extend with other menus

They share a common layout, and the URL context menu changes the title
text size programmatically to display the potentially longer text. There's
a new subheading view that's collapsed by default, and set on the URL
menu when showing GIF sizes. The padding etc has been changed to fit the
material dialog guidelines

The fragments also save and restore state, so if the menu gets recreated
(say it's displayed and the user minimises the app for long enough that
the activity is recreated) then it still works as expected and avoids a
crash.

I've also refactored some variables that were getting passed around and
cast back and forth between integers and strings. Now they're handled
consistently as integers, and only converted to strings when required
(e.g. as a parameter in a request)