Add favourite forums
I've added a new toggleable forum list that can also display a customisable list
of favourites. The user can add/remove favourites by long-pressing a forum,
or using the 'manage forums' button in the favourites view to edit in bulk.

Favourites are stored as a SharedPreferences item for persistence and so they'll
survive backups. Right now there's no way to re-order forums, they're just
presented in the order they appear in the main list - this wouldn't be hard to add
on the backend, but it needs some UI to handle moving them around.

I've had to rework ForumsIndexFragment a little to repurpose the empty forums
view (so it can be used for 'no favourites') and make the title change to reflect
the current mode (all forums or just favourites).

I've also had to mess with ForumStructure a little to allow for a hierarchy that
doesn't ultimately share the same root forum ID, since favourites can
be any nodes in the tree really. That whole class is a bit confusing and
the attempts at immutability for Forums is not really helping in Java.
Good candidate for Kotlin I reckon!