Add Imgur uploader/inserter
Shows a dialog where the user can select an image / add a URL and have
it hosted on Imgur, automatically inserting the hosted image as BBcode
in the message composer.

The app needs an API key for Imgur - I've added a secrets.xml file to
the .gitignore to add the key as a resource that's not checked in. The
free tier puts credit limits on the app and blocks it if too many
requests are made, so the ImgurUploadRequest class keeps track of the
current limits and enforces a per-user daily quota (to stop a few users
using all the app's credits)

The credit limits might need tweaking for a better experience, depending
on how many people actually use it and how much. I just guessed at some
limits!