Rewrite thread page and thread list parsing
There were some weird things happening in there, assumptions about post
numbers and some logic that was really hard to follow.

I've tried to clean it up and make it more understandable, and also
rewrote some of the selector logic to cut down the steps needed to grab
the data.

There are still some issues, mainly that ignored posts aren't counted,
and the "show user's posts" thing is... crazy
It writes over the thread records, starting at the beginning, and
updates the metadata to reflect that cut-down view. So the read/unread
count changes, and a lot of posts get overwritten until that page is
reloaded in the 'full' thread view (not filtering by user)

This is also why you get random other posts on a page with a user's
posts - if they have enough to fill half a page, that overwrites half
of page 1, and then when the thread view loads that page there are other
posts on that page, because the entire rest of the thread is still there