Rework AwfulThread as data class, add de/serialisation, and use it
I've added data fields to the class so instances can actually be used
as a representation of a thread, instead of messing around with database
conversions in unrelated code. There are methods to generate
an AwfulThread instance from a Cursor row, and to turn one back into a
set of ContentValues.

This means we can just have other code reference the thread's fields,
and it'll be easier to modify the thread data structure and pass the
object around instead of a bunch of parameters. Ideally we can abstract
everything away, so we can fix things like the thread view monitoring a
literal page of database records for a fixed thread ID

The whole Cursor/ContentValues conversion thing is a holdover since
that's how things basically work at the moment, once things are separated
it will probably make more sense to have a ThreadRepository that stores
and gets sets of threads, and handles all the query stuff internally