Tweak Imgur dialog layout, add error handling/validation and GIF options
I've moved some things around on the layout to hopefully make it a little less
dense, including a circle spinner. The dialog is now scrollable in case it
doesn't fit on the device's screen (e.g. if you're using it in landscape for
some reason). The scrolling is not ideal at all, since it can leave
the progress bar/status out of view - I wasn't able to put those things at
the top without the layout breaking/collapsing/other stupid behaviour. It really
needs to be in a fullscreen dialog when space is limited, especially since it's
a fairly complicated dialog and might get even more added to it

I've added better error handling, and there's some basic validation on the URL
box (the 'looks like a URL' check is only advisory since it's not exactly robust).
There's also an option to embed the video version of a GIF, if there is one.
Since that involves passing the URL to the UrlInserter, and the same change
had to be made to the ImageInserter, I just did it to the whole set of Inserter
classes so they're consistent.

Debug builds now get less restricted credits (the app limit still has to apply!)