Fix Crashlytics #774, some cleanup in ForumsIndexActivity
A fragment was getting activity-requiring methods called on it during the
New Intent > Show Pager Fragment process, and the fragment wasn't actually
attached so these were crashing. I couldn't replicate it at all, but I added a check
so they aren't called (they should perform title updates themselves when ready)

Also removed some magic numbers, and redid the fragment-storing code
(everyone recommends avoiding it in #getItem because it's not always called,
e.g. if the fragments are already cached in the fragment manager)