Create StatusFrog widget, add to layouts using something similar
I've recreated the frog icon/message/progress spinner that's used in various
places (empty forum list, empty message pane in the PMs tablet layout) and
turned it into a drop-in widget. You can add it in XML, set the message and
spinner status, and then tweak those programmatically to update the current
status

That's been added to the Announcements fragment, so now it shows status
for fetching, failure and 'no announcements' instead of just being a blank page. I've also
added it to other layouts that already used something similar, replacing the existing views
and updating the code that manipulates them