Add Instagram embedding, create separate JS file for embedding content
I've added an embedding function and for Instagram URLs, and a preference that
defaults to on (for discovery).

I put the JS function in a separate script file where all the embedding code
can go, just to make it easier to extend without complicating the general thread
setup code. So now there's just a single call to the embed script, and that runs
everything according to the preferences that are passed in.

Also added some Zepto libraries to enable Promises - handy for things like the
Instagram oEmbed code, where you need to add HTML for each element
and then process the finished page when they're all done.