Fix SSL connection failures on pre-Lollipop, also updates network security
This uses the security provider from Google Play Services to basically patch
the networking calls and stop old devices from trying to connect with SSLv3,
which doesn't work anymore. It should also make things more secure in general?

I'm using an old version of the auth library so I can test it on an old emulator.
Might be worth updating later? You need Google Play Services installed in the SDK
too

I've crowbarred an SSL error check into the base request class, seems like the
easiest place to catch it. It just pops up a toast telling people to update
Google Play Services so the security provider can hopefully update. Feels bad
but it's a fix.

I'm working on a system for popping a useful info dialog but I'm not sure how to
tell when it's important ("you need to update because SSL is broke as heck")
or not ("you need to update for better security even though everything works").
And people *really don't like popups*