Add gradle task for updating thread tags, and update them in the app
I've added a task to automatically copy PNGs from a source folder, following
renaming rules (rewriting hyphens to underscores and using a different format
for ratings stars). Just run it as a commandline gradle task (the Execute Gradle
Task button on the Gradle panel is easiest) and pass the path to the folder that
contains the tag PNGs. It ignores every other file, so you can just point it at
your local copy of the tag repo

It doesn't delete anything since it's dumping into the drawable-ldpi folder -
we could have a dedicated tag folder in assets and wipe that clean each time,
but that's more complicated than just accessing drawables

I've run the update and added the latest tags too