Add timestamps to forum search results, change forum filter icon, some cleanup
I've added the timestamp data to the search results - just as an extra line
for now, maybe the layout can be reworked

While I was at it I converted the result list items to use a ConstraintLayout,
and added the list divider used in other RecyclerViews (like the forum list).
I also fixed the padding in the whole activity to bring it more in line with the
rest of the app

Also fixed the title, the tooltip on the forum filter icon (it said "Submit")
and changed the icon from the settings gear to the Material filter one. Reworded
a couple of things and added string resources, and fixed some basic lint warnings
(still some important ones outstanding)