Revert changes to ViewPager state handling in ForumsIndexActivity
Ideally the ViewPager should be handling and storing its own pages internally,
without other classes reaching in and calling getItem() - that method is just
meant to generate new pages on demand, but ours has some added caching behaviour.

Removing that breaks things like volume scrolling, which ends up creating a new
page fragment and calling methods on it immediately, before it's initialised. So
I'm reverting those changes to avoid the crash, but the pager should probably
be reworked into its own separate class. (Which is complex because there's a
whole lot of weird behaviour holding everything together, so we need a quick fix)