Add PostPreview parser to ForumParsing (fixes preview errors)
I've made a separate, simple task for handling previews. Previously it was
trying to parse an entire post from the element (even though most of the stuff
is missing from the preview section), and then pulling the HTML from the resulting
AwfulPost object. Now it just extracts the HTML directly and returns that.