Adjust min width of page counter to keep button areas stable
I've added a large minimum width so a large readout ("Page 9999/9999") doesn't
exceed it. This means the areas for the icons *shouldn't* grow and shrink,
making the layout look fixed - the downside is that with a smaller readout,
the icons look more offset (because some of the space belongs to the page
counter)

So the tradeoff is between a layout where the icons can slide around ("dynamic"!)
vs one where the space doesn't look evenly split. We'll see how it goes, maybe
a smaller minWidth is a better compromise