Update PageBar class with better state handling and transitions
I've added some explicit states to reflect what the page bar is displaying -
basically whether it's a single page, or the first or last, or a normal middle
page, or one where we don't know how many pages there are. This way it's clearer
about defining what we need to show in each state, and how we show it.

I've also redone the code to show and hide the various elements - by doing the
hiding first, things disappear cleanly and get replaced, instead of things
appearing and overlapping before the other element disappears.

I want to redo this in Kotlin because I feel like it'll be a lot more concise