Extract NavigationDrawer to separate class, cleanup in ForumsIndexActivity
The navigation drawer is now its own separate component added to the activity.
It handles its own display, listens for relevant callbacks (e.g. new
announcements) and calls the activity when the user clicks on an item.

I've refactored the activity so it exposes methods for these navigation events -
so now there are functions for "show announcements", "show logout" etc. This
basically extends the interface we were already using ("display thread", "display
forum") and makes the activity more of a central coordinator for any events one
of the components might trigger.

It's still not perfect, but until the rest of the activity/fragment code and
interaction is cleaned up, it at least works like it used to

I also cleaned up a few lint issues while I was at it (adding lambdas mostly)

Docs