Refactor openForum/Thread code to make page numbers optional
Now you can ask the forum and thread fragments to just display something without
specifying a page. This lets us basically tell the view pager to show a thread
or forum, and the fragment can decide whether it needs to reload or not. Keeps
all the current state and update logic within the fragment.