Fix ForumsIndexActivity titles
Mimics old behaviour - just shows the title of the currently selected page
(the one on the left, in tablet mode). There's code there to display a combined
title for when multiple fragments are visible, but it seems to be impossible to
actually work out what's showing and what isn't. Might need to write a whole
bunch of custom state tracking code, thanks Android

On the plus side fragments don't push their titles to the action bar anymore -
the pager controller tells the Activity when it's showing different pages, and fragments
can notify the activity that their contents have changed. The activity then
asks the viewpager controller for a current title (since it knows what's showing
and what isn't, or it *should*)