Fix viewpager initialisation, simplify pager controller structure
Reworked the pager update callback system, so when pages are initialised or
restored they trigger an update so the activity can poll for a title.

(Before this wasn't happening because #onPageSelected was firing before we had
the fragments, so there was nothing to poll for a title. The fragments updating
themselves caused a separate callback to the activity, but in a restore situation
this was basically a race condition, and sometimes the fragments would do their
callbacks before the viewpager had been restored with access to them)

Right now there are multiple calls to setActionbarTitle - this is basically down
to the various events that get triggered (e.g. a restored fragment notifies the
activity its contents have changed, the viewpager selects a page but it doesn't
have the fragments yet, the fragment is added and it's the current page so it
notifies the activity). Each one is really a separate event so it's probably
cleanest to just let it happen - it will always end up with the title of the
current page. Some calls are down to repeated "I've updated" calls from the
fragments themselves, so those can be cleaned up when the fragments are reworked

Also simplified some code