Rework main activity, page handling, events and navigation (#583)
A lot of the code in ForumsIndexActivity has been broken out into other components. There's a viewpager controller that basically manages the forum pages, what's being displayed, and events like "show this page" or "open this thread" (which also involves showing a page). It handles restoration of its own state (including getting references to the existing fragments when it's recreated) so all that complexity is kept in one place. There's still some stuff exposed for the moment (the activity can still get to the fragments) but that might get refined later. It also handles generating a title for the activity, so instead of fragments fighting to change it, the controller can work out what's being shown and produce a title.

Fragments handle their own state (not fully working yet), and are basically supposed to look after themselves - display something, restore when necessary, let the activity know when they've changed so it can react to that event. The activity is mostly a coordinator at this point, passing events around and providing a navigation interface - there are methods like showThread, showBookmarks etc that anything can trigger

That interface has also been pushed up into AwfulActivity, which basically allows any activity in the app to fire those commands. All the intent handling is done there - there's a new NavigationEvent type that basically defines things like "show this thread" or "show the forum list", that vastly simplifies the onNewIntent code

It should be easier to work with, and fixes a lot of bugs, conflicts between fragments trying to set titles and the like. Oh and you should be able to put the NavigationDrawer into any AwfulActivity now and have it Just Work (haven't tried it yet though....)